backend/requirements.txt
fastapi
pydantic
pyjwt
uvicorn
SQLAlchemy==2.0.40
psycopg2-binary
geoalchemy2
passlib
